+++ source/extensions/lightspeed.trex.stage_manager.widget/lightspeed/trex/stage_manager/widget/stage_manager.py
@@ -1,0 +1,25 @@
+"""
+* SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+* SPDX-License-Identifier: Apache-2.0
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* https://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+"""
+
+from omni.flux.stage_manager.widget import StageManagerWidget as _StageManagerWidget
+
+
+class StageManagerWidget(_StageManagerWidget):
+    # This extension is only used to:
+    # - Pull all the plugins required for the Lightspeed StageManager Widget
+    # - Define the schema path in the settings
+    pass