+++ source/extensions/omni.flux.stage_manager.factory/omni/flux/stage_manager/factory/plugins/widget_plugin.py
@@ -16,7 +16,11 @@
 """
 
 import abc
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Callable
+
+from omni.flux.utils.common import Event as _Event
+from omni.flux.utils.common import EventSubscription as _EventSubscription
+from pydantic import PrivateAttr
 
 from .base import StageManagerUIPluginBase as _StageManagerUIPluginBase
 
@@ -30,6 +34,8 @@ class StageManagerWidgetPlugin(_StageManagerUIPluginBase, abc.ABC):
     A plugin that provides a widget for the TreeView
     """
 
+    _on_item_clicked: _Event = PrivateAttr(_Event())
+
     @abc.abstractmethod
     def build_ui(  # noqa PLW0221
         self, model: "_StageManagerTreeModel", item: "_StageManagerTreeItem", level: int, expanded: bool
@@ -39,3 +45,31 @@ class StageManagerWidgetPlugin(_StageManagerUIPluginBase, abc.ABC):
     @abc.abstractmethod
     def build_overview_ui(self, model: "_StageManagerTreeModel"):
         pass
+
+    def subscribe_item_clicked(
+        self, callback: Callable[[int, bool, "_StageManagerTreeModel", "_StageManagerTreeItem"], None]
+    ) -> _EventSubscription:
+        """
+        Subscribe to the event that is triggered when a widget is clicked.
+
+        Args:
+            callback: A function that will be called when the widget is clicked.
+
+        Returns:
+            An object that will automatically unsubscribe when destroyed.
+        """
+        return _EventSubscription(self._on_item_clicked, callback)
+
+    def _item_clicked(
+        self, button: int, should_validate: bool, model: "_StageManagerTreeModel", item: "_StageManagerTreeItem"
+    ):
+        """
+        Callback called whenever a widget is clicked on.
+
+        Args:
+            button: The mouse button that triggered the event
+            should_validate: Whether the TreeView selection should be validated or not
+            model: The tree model
+            item: The tree item that was clicked
+        """
+        self._on_item_clicked(button, should_validate, model, item)