+++ source/extensions/lightspeed.event.shutdown/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.5]
+### Fixed
+- Fixed tests
+
 ## [1.0.4]
 ### Added
 - Added a test for unload stage save prompt