+++ source/extensions/lightspeed.trex.app.resources/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.6.9]
+### Changed
+- Updated `default_schema` widget to use renamed `IsVisibleActionWidgetPlugin`
+
 ## [1.6.8]
 ### Added
 - Another image for testing