+++ source/extensions/lightspeed.trex.control.stagecraft/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.1]
+### Fixed
+- Fixed tests
+
 ## [1.2.0] - 2024-08-13
 ### Added
 - Added `_previous_root_layer_identifier` and `_on_reload_last_workfile()` for reloading previous stages